extracted LCOVCoverageSensor
